[svn] ConEmu sources
git-svn-id: https://conemu-maximus5.googlecode.com/svn@1150 92116c04-ab5e-ca4c-3a8b-6f4eadf4b4a1

+++ src/ConEmu/TabID.h
@@ -138,7 +138,7 @@ struct TabInfo
 class CTabID
 {
 protected:
-	int mn_RefCount;
+	LONG mn_RefCount;
 	~CTabID();
 public:
 	TabInfo Info;
